This import represents the merge of Andrea Scarpelli's feature branch (feature/ascarpel_timing_pmt_crt, see Pull Request #440) with updates to fhicl files to include the CRT hits in stage 0. This became a bit complicated by the base of Andrea's feature branch being before the restructuring of fhicl files so I've had to reconcile that. As well, I moved the new PMT data product to the IcarusObj folder so we can be sure of where are data objects are. Note that I did not reconcile the fhicl files directly related to production running so some of the special case PMT fhicl files still need reconciliation.
