+++ fcl/reco/Definitions/stage0_icarus_defs.fcl
@@ -8,10 +8,12 @@
 #include "decoderdefs_icarus.fcl"
 #include "recowire_icarus.fcl"
 #include "hitfindermodules_icarus.fcl"
+#include "timing_icarus.fcl"
 #include "icarus_ophitfinder.fcl"
 #include "icarus_flashfinder.fcl"
 #include "trigger_emulation_icarus.fcl"
 #include "crt_decoderdefs_icarus.fcl"
+#include "crthitproducer.fcl"
 #include "wcls-decode-to-sig-base.fcl"
 
 BEGIN_PROLOG
@@ -28,6 +30,19 @@ icarus_stage0_analyzers:
                     PurityInfoLabel: "purityana1"
                     PrintInfo:       false
                     SelectEvents:    [ reco ]
+                  }
+  CRTPMTMatching: { module_type: "CRTPMTMatchingAna"
+                    OpHitModuleLabel:    "ophit"
+                    OpFlashModuleLabel0: "opflashCryoE"
+                    OpFlashModuleLabel1: "opflashCryoW"
+                    CrtHitModuleLabel:   "crthit"
+                    TriggerLabel:        "daqTrigger" 
+                    FlashPeThresh:       2000.0
+                    HitPeThresh:         100.0
+                    CoincidenceWindow:   80.0
+                    OpDelay:             43.0
+                    HitVelocityMax:      15.0
+                    HitVelocityMin:      2.0
                   }
 }
 
@@ -71,7 +86,7 @@ icarus_stage0_producers:
   roifinder:                      @local::icarus_roifinder
   roifinder2d:                    @local::icarus_roifinder
 
-  ### hit-finder producers
+  ### TPC hit-finder producers
   gaushitTPCWW:                   @local::gaus_hitfinder_icarus
   gaushitTPCWE:                   @local::gaus_hitfinder_icarus
   gaushitTPCEW:                   @local::gaus_hitfinder_icarus
@@ -82,13 +97,18 @@ icarus_stage0_producers:
   gaushit2dTPCEW:                 @local::gaus_hitfinder_icarus
   gaushit2dTPCEE:                 @local::gaus_hitfinder_icarus
 
+  ### CRT hit finder producer
+  crthit:                         @local::standard_crthitproducer	
+
   ### trigger emulation foundation
   pmtconfigbaselines:             @local::icarus_pmtconfigbaselines
   pmtthr:                         @local::icarus_pmtdiscriminatethr
   
   ### Optical hit finder
-  ophit:                          @local::icarus_ophit_data
-  ophitfull:                      @local::icarus_ophitdebugger_data
+  ophituncorrected:               @local::icarus_ophit_data
+  ophit:                          @local::icarus_ophit_timing_correction
+  ophitfulluncorrected:           @local::icarus_ophitdebugger_data
+  ophitfull:                      @local::icarus_ophit_timing_correction
   opflashCryoE:                   @local::ICARUSSimpleFlashDataCryoE
   opflashCryoW:                   @local::ICARUSSimpleFlashDataCryoW
 
@@ -214,6 +234,7 @@ icarus_stage0_PMT:                 [ triggerconfig,
                                      daqPMT,
                                      pmtconfigbaselines,
                                      pmtthr,
+                                     ophituncorrected,
                                      ophit,
                                      opflashCryoE,
                                      opflashCryoW
@@ -241,22 +262,24 @@ icarus_stage0_2d_multiTPC:         [ @sequence::icarus_stage0_multiTPC_2d_TPC,
                                      @sequence::icarus_purity_monitor
                                    ]
 
-icarus_stage0_crt:                 [
-                                     daqCRT
+icarus_stage0_CRT:                 [
+                                     daqCRT,
+                                     crthit
                                    ]
 
 icarus_stage0_data:                [
                                      @sequence::icarus_stage0_PMT, 
                                      daqTPCROI, 
                                      @sequence::icarus_stage0_multiTPC,
-                                     @sequence::icarus_stage0_crt
+                                     @sequence::icarus_stage0_CRT
                                    ]
 
 icarus_stage0_2d_data:             [
                                      @sequence::icarus_stage0_PMT, 
+                                     @sequence::icarus_stage0_CRT,
                                      daqTPCROI, 
                                      @sequence::icarus_stage0_2d_multiTPC,
-                                     @sequence::icarus_stage0_crt
+                                     @sequence::icarus_stage0_CRT
                                    ]
 
 ### Below we include overrides for the modules above
@@ -409,9 +432,11 @@ icarus_stage0_producers.gaushit2dTPCEE.HitFinderToolVec.CandidateHitsPlane2:    
 icarus_stage0_producers.gaushit2dTPCEE.HitFinderToolVec.CandidateHitsPlane2.Plane:             2
 icarus_stage0_producers.gaushit2dTPCEE.HitFinderToolVec.CandidateHitsPlane2.RoiThreshold:      9.
 
-### Optical hit finding
-icarus_stage0_producers.ophit.InputModule:                                                     "daqPMT"
-icarus_stage0_producers.ophit.InputLabels:                                                     []
+###
+### Optical detector
+###
+icarus_stage0_producers.ophit.InputLabels:                                                     [ "ophituncorrected" ]
+icarus_stage0_producers.ophitfull.InputLabels:                                                 [ "ophitfulluncorrected" ]
 
 ### Default purity monitor settings (single TPC readout assumed)
 icarus_stage0_producers.purityana0.RawModuleLabel:                                             ["daqTPCROI:PHYSCRATEDATATPCWW","daqTPCROI:PHYSCRATEDATATPCWE","daqTPCROI:PHYSCRATEDATATPCEW","daqTPCROI:PHYSCRATEDATATPCEE"]
