+++ fcl/reco/Stage0/Run2/stage0_run2_icarus.fcl
@@ -18,7 +18,13 @@ physics.end_paths:     [ outana, streamROOT ]
 # Drop all output from the TPC decoder stage
 # Drop all output from the 1D deconvolution stage
 # Drop the recob::Wire output from the roifinder (but keep the ChannelROIs)
-outputs.rootOutput.outputCommands:         ["keep *_*_*_*", "drop *_*_*_DAQ*", "drop *_daqTPCROI_*_*", "drop *_decon1droi_*_*", "drop recob::Wire*_roifinder_*_*" ]
+outputs.rootOutput.outputCommands:         ["keep *_*_*_*", 
+                                            "drop *_*_*_DAQ*", 
+                                            "drop *_ophituncorrected_*_*",
+                                            "drop *_daqTPCROI_*_*", 
+                                            "drop *_decon1droi_*_*", 
+                                            "drop recob::Wire*_roifinder_*_*" 
+                                           ]
 
 ## Modify the event selection for the purity analyzers
 physics.analyzers.purityinfoana0.SelectEvents:    [ path ]
