+++ fcl/reco/Stage1/Run2/stage1_run2_icarus.fcl
@@ -5,7 +5,7 @@ process_name: stage1
 physics.reco: [ flashfilter,
                 @sequence::icarus_filter_cluster3D,
                 @sequence::icarus_pandora_Gauss,
-                @sequence::icarus_reco_fm,
+                #@sequence::icarus_reco_fm,
                 @sequence::icarus_crthit,
                 @sequence::icarus_crtt0match,
                 caloskimCalorimetryCryoE, caloskimCalorimetryCryoW]
