+++ icaruscode/CRT/CRTPMTMatchingAna_module.cc
@@ -219,265 +219,273 @@ icarus::crt::CRTPMTMatchingAna::CRTPMTMatchingAna(fhicl::ParameterSet const& p)
 
 void icarus::crt::CRTPMTMatchingAna::analyze(art::Event const& e)
 {
-  // Implementation of required member function here.
-  /*
-  geo::CryostatGeo const& cryo0 = fGeometryService->Cryostat(0);
-  geo::CryostatGeo const& cryo1 = fGeometryService->Cryostat(1);
-  geo::TPCGeo const& tpc00 = cryo0.TPC(0);
-  geo::TPCGeo const& tpc01 = cryo0.TPC(1);
-  geo::TPCGeo const& tpc10 = cryo1.TPC(0);
-  geo::TPCGeo const& tpc11 = cryo1.TPC(1);
-  */
-  MF_LOG_DEBUG("CRTPMTMatching: ") << "beginning analyis" << '\n';
+    // Implementation of required member function here.
+    /*
+    geo::CryostatGeo const& cryo0 = fGeometryService->Cryostat(0);
+    geo::CryostatGeo const& cryo1 = fGeometryService->Cryostat(1);
+    geo::TPCGeo const& tpc00 = cryo0.TPC(0);
+    geo::TPCGeo const& tpc01 = cryo0.TPC(1);
+    geo::TPCGeo const& tpc10 = cryo1.TPC(0);
+    geo::TPCGeo const& tpc11 = cryo1.TPC(1);
+    */
+    MF_LOG_DEBUG("CRTPMTMatching: ") << "beginning analyis" << '\n';
 
-  // Start by fetching some basic event information for our n-tuple.
-  fEvent  = e.id().event();
-  fRun    = e.run();
-  fSubRun = e.subRun();
+    // Start by fetching some basic event information for our n-tuple.
+    fEvent  = e.id().event();
+    fRun    = e.run();
+    fSubRun = e.subRun();
 
-  ClearVecs();
+    ClearVecs();
 
-  //add trigger info
-  if( !fTriggerLabel.empty() ) {
+    //add trigger info
+    if( !fTriggerLabel.empty() ) 
+    {
+        art::Handle<sbn::ExtraTriggerInfo> trigger_handle;
+        e.getByLabel( fTriggerLabel, trigger_handle );
+        if( trigger_handle.isValid() ) 
+        {
+            sbn::triggerSource bit = trigger_handle->sourceType;
+            m_gate_type = (unsigned int)bit;
+            m_gate_name = bitName(bit);
+            m_trigger_timestamp = trigger_handle->triggerTimestamp;
+            m_gate_start_timestamp =  trigger_handle->beamGateTimestamp;
+            m_trigger_gate_diff = trigger_handle->triggerTimestamp - trigger_handle->beamGateTimestamp;
 
-    art::Handle<sbn::ExtraTriggerInfo> trigger_handle;
-    e.getByLabel( fTriggerLabel, trigger_handle );
-    if( trigger_handle.isValid() ) {
-      sbn::triggerSource bit = trigger_handle->sourceType;
-      m_gate_type = (unsigned int)bit;
-      m_gate_name = bitName(bit);
-      m_trigger_timestamp = trigger_handle->triggerTimestamp;
-      m_gate_start_timestamp =  trigger_handle->beamGateTimestamp;
-      m_trigger_gate_diff = trigger_handle->triggerTimestamp - trigger_handle->beamGateTimestamp;
-
+        }
+        else{
+            mf::LogError("CRTPMTMatching:") << "No raw::Trigger associated to label: " << fTriggerLabel.label() << "\n" ;
+        }
     }
-    else{
-      mf::LogError("CRTPMTMatching:") << "No raw::Trigger associated to label: " << fTriggerLabel.label() << "\n" ;
-    }
-  }
-  else {
-    std::cout  << "Trigger Data product " << fTriggerLabel.label() << " not found!\n" ;
-  }
-
-
-  //OpHits
-  art::Handle< std::vector<recob::OpHit> > opHitListHandle;
-  std::vector< art::Ptr<recob::OpHit> >    opHitList;
-  if( e.getByLabel(fOpHitModuleLabel,opHitListHandle) )
-    art::fill_ptr_vector(opHitList, opHitListHandle);
-
-
-  //OpFlash
-  map<int, art::Handle< std::vector<recob::OpFlash> > > flashHandles;
-  std::map<int,std::vector< art::Ptr<recob::OpFlash> >> opFlashLists;
-  //  fNFlash = 0;
-  for(int i=0; i<2; i++) {
-    if( e.getByLabel(fFlashLabels[i],flashHandles[i]) )
-      art::fill_ptr_vector(opFlashLists[i], flashHandles[i]);
-  }
-
-
-  //CRTHits
-  art::Handle< std::vector<CRTHit> > crtHitListHandle;
-  std::vector< art::Ptr<CRTHit> >    crtHitList;
-  if( e.getByLabel(fCrtHitModuleLabel,crtHitListHandle))
-    art::fill_ptr_vector(crtHitList, crtHitListHandle);
-
-  fNCrt = crtHitList.size();
-
-  for(auto const& crt : crtHitList){
-    vector<double> xyzt, xyzerr;
-    TVector3 rcrt(crt->x_pos,crt->y_pos,crt->z_pos);
-    m_gate_crt_diff = m_gate_start_timestamp - crt->ts0_ns;
-    xyzt.push_back(rcrt.X());
-    xyzt.push_back(rcrt.Y());
-    xyzt.push_back(rcrt.Z());
-
-    double tcrt = double(m_gate_start_timestamp - crt->ts0_ns)/1e3;
-    tcrt = -tcrt+1e6;
-
-    //    double tcrt = (int32_t)crt->ts0_ns;// - fCrtDelay;
-    //uint64_t tcrt = crt->ts0_ns;
-
-    //    std::cout <<"T0: " << crt->ts0_ns << " , tcrt : " << tcrt << std::endl;
-    xyzt.push_back(tcrt);
-    fCrtXYZT.push_back(xyzt);
-
-    xyzerr.push_back(crt->x_err);
-    xyzerr.push_back(crt->y_err);
-    xyzerr.push_back(crt->z_err);
-    fCrtXYZErr.push_back(xyzerr);
-
-    fCrtPE.push_back(crt->peshit);
-    fCrtRegion.push_back(crtutil->AuxDetRegionNameToNum(crt->tagger));
-
-    // -- flash match --
-    int matchtpc = -1;
-    double tdiff = DBL_MAX, rdiff=DBL_MAX, peflash=DBL_MAX;
-    bool matched=false;
-    xyzt.clear();
-    double flashHitT = DBL_MAX, flashHitPE=DBL_MAX, flashHitDiff=DBL_MAX;
-    vector<double> flashHitxyzt;
-
-    for(auto const& flashList : opFlashLists) {
-
-      art::FindManyP<recob::OpHit> findManyHits(flashHandles[flashList.first], e, fFlashLabels[flashList.first]);
-
-      for(size_t iflash=0; iflash<flashList.second.size(); iflash++) {
-
-	auto const& flash = flashList.second[iflash];
-	if(flash->TotalPE()<fFlashPeThresh){
-	  continue;
-	}
-
-	double tflash = flash->Time();//*1e3;//-fOpDelay;
-
-	TVector3 rflash(0,flash->YCenter(),flash->ZCenter());
-	TVector3 vdiff = rcrt-rflash;
-	//std::cout << "flash time: "<<flash->Time() << " absdiff : "<< abs(tcrt-tflash)<< std::endl;
-	if(abs(tcrt-tflash)<abs(tdiff)) {
-	  peflash = flash->TotalPE();
-	  tdiff = tcrt-tflash;
-	  rdiff = vdiff.Mag();
-	  xyzt.clear();
-	  xyzt.push_back(rflash.X());
-	  xyzt.push_back(rflash.Y());
-	  xyzt.push_back(rflash.Z());
-	  xyzt.push_back(tflash);
-	  matched = true;
-	  matchtpc = flashList.first;
-
-	  vector<art::Ptr<recob::OpHit>> hits = findManyHits.at(iflash);
-	  for(auto const& hit : hits) {
-	    double tPmt = hit->PeakTime();//*1.e3;//-fOpDelay;
-	    if( tPmt < flashHitT) {
-	      flashHitT = tPmt;
-	      flashHitPE = hit->PE();
-	      //FlashHit position/time
-	      double pos[3];
-	      fGeometryService->OpDetGeoFromOpChannel(hit->OpChannel()).GetCenter(pos);
-	      flashHitxyzt.clear();
-	      for(int i=0; i<3; i++) flashHitxyzt.push_back(pos[i]);
-	      flashHitxyzt.push_back(flashHitT);
-
-	      //FlashHit distance
-	      TVector3 rflashHit(pos[0],pos[1],pos[2]);
-	      TVector3 vdiffHit = rcrt-rflashHit;
-	      flashHitDiff = vdiffHit.Mag();
-	    }
-	  }//loop over flash hits
-	}//if minimum tdiff
-      }//for OpFlash in this flash list
-    }//for flash lists
-    if(!matched) {
-      peflash = DBL_MAX;
-      for(int i=0; i<4; i++) xyzt.push_back(DBL_MAX);
+    else {
+        std::cout  << "Trigger Data product " << fTriggerLabel.label() << " not found!\n" ;
     }
 
-    fMatchFlash.push_back(matched);
-    fTofFlash.push_back(tdiff);
-    fTofPeFlash.push_back(peflash);
-    fTofXYZTFlash.push_back(xyzt);
-    fDistFlash.push_back(rdiff);
-    fTofTpcFlash.push_back(matchtpc);
-    fTofFlashHit.push_back(tcrt-flashHitT);
-    fTofPeFlashHit.push_back(flashHitPE);
-    fTofXYZTFlashHit.push_back(flashHitxyzt);
-    fDistFlashHit.push_back(flashHitDiff);
-
-    // -- match OpHits to CRTHits --
-    tdiff = DBL_MAX;
-    matched = false;
-    peflash = DBL_MAX;
-    rdiff = DBL_MAX;
-    xyzt.clear();
-
-    double pemax = 0.;
-    for(auto const& hit : opHitList) {
-      double thit = hit->PeakTime();//*1e3-fOpDelay;
-      if(hit->PE()<fHitPeThresh){
-	continue;
-      }
+    //OpHits
+    art::Handle< std::vector<recob::OpHit> > opHitListHandle;
+    std::vector< art::Ptr<recob::OpHit> >    opHitList;
+    if( e.getByLabel(fOpHitModuleLabel,opHitListHandle) )
+        art::fill_ptr_vector(opHitList, opHitListHandle);
 
 
-      //if(abs(tcrt-thit)<abs(tdiff)) {
-      if(abs(tcrt-thit)<fCoinWindow && hit->PE()>pemax) {
-	pemax = hit->PE();
-
-	//hitXYZT
-	double pos[3];
-	fGeometryService->OpDetGeoFromOpChannel(hit->OpChannel()).GetCenter(pos);
-
-	//distHit
-	TVector3 rhit (pos[0],pos[1],pos[2]);
-	TVector3 vdiff = rcrt-rhit;
-	rdiff = vdiff.Mag();
-	peflash = hit->PE();
-	tdiff = tcrt-thit;
-	xyzt.clear();
-	for(int i=0; i<3; i++) xyzt.push_back(pos[i]);
-	xyzt.push_back(thit);
-
-	matched = true;
-	std::cout << "thit: "<<thit << " , tcrt: " << tcrt << " , tdiff " << tdiff  << std::endl;
-      }//if min tof
-    }//for OpHits
-    if(!matched) {
-      tdiff = DBL_MAX;
-      peflash = DBL_MAX;
-      rdiff = DBL_MAX;
-      xyzt.clear();
-      for(int i=0; i<4; i++) xyzt.push_back(DBL_MAX);
+    //OpFlash
+    map<int, art::Handle< std::vector<recob::OpFlash> > > flashHandles;
+    std::map<int,std::vector< art::Ptr<recob::OpFlash> >> opFlashLists;
+    //  fNFlash = 0;
+    for(int i=0; i<2; i++) {
+        if( e.getByLabel(fFlashLabels[i],flashHandles[i]) )
+            art::fill_ptr_vector(opFlashLists[i], flashHandles[i]);
     }
-    fMatchHit.push_back(matched);
-    fTofHit.push_back(tdiff);
-    fTofPeHit.push_back(peflash);
-    fTofXYZTHit.push_back(xyzt);
-    fDistHit.push_back(rdiff);
-    fTofTpcHit.push_back(matchtpc);
-  }//for CRTHits
 
-  fMatchTree->Fill();
+
+    //CRTHits
+    art::Handle< std::vector<CRTHit> > crtHitListHandle;
+    std::vector< art::Ptr<CRTHit> >    crtHitList;
+    if( e.getByLabel(fCrtHitModuleLabel,crtHitListHandle))
+        art::fill_ptr_vector(crtHitList, crtHitListHandle);
+
+    fNCrt = crtHitList.size();
+
+    for(auto const& crt : crtHitList)
+    {
+        vector<double> xyzt, xyzerr;
+        TVector3 rcrt(crt->x_pos,crt->y_pos,crt->z_pos);
+        m_gate_crt_diff = m_gate_start_timestamp - crt->ts0_ns;
+        xyzt.push_back(rcrt.X());
+        xyzt.push_back(rcrt.Y());
+        xyzt.push_back(rcrt.Z());
+
+        double tcrt = double(m_gate_start_timestamp - crt->ts0_ns)/1e3;
+        tcrt = -tcrt+1e6;
+
+        //    double tcrt = (int32_t)crt->ts0_ns;// - fCrtDelay;
+        //uint64_t tcrt = crt->ts0_ns;
+
+        //    std::cout <<"T0: " << crt->ts0_ns << " , tcrt : " << tcrt << std::endl;
+        xyzt.push_back(tcrt);
+        fCrtXYZT.push_back(xyzt);
+
+        xyzerr.push_back(crt->x_err);
+        xyzerr.push_back(crt->y_err);
+        xyzerr.push_back(crt->z_err);
+        fCrtXYZErr.push_back(xyzerr);
+
+        fCrtPE.push_back(crt->peshit);
+        fCrtRegion.push_back(crtutil->AuxDetRegionNameToNum(crt->tagger));
+
+        // -- flash match --
+        int matchtpc = -1;
+        double tdiff = DBL_MAX, rdiff=DBL_MAX, peflash=DBL_MAX;
+        bool matched=false;
+        xyzt.clear();
+        double flashHitT = DBL_MAX, flashHitPE=DBL_MAX, flashHitDiff=DBL_MAX;
+        vector<double> flashHitxyzt;
+
+        for(auto const& flashList : opFlashLists) 
+        {
+            art::FindManyP<recob::OpHit> findManyHits(flashHandles[flashList.first], e, fFlashLabels[flashList.first]);
+
+            for(size_t iflash=0; iflash<flashList.second.size(); iflash++) 
+            {
+	            auto const& flash = flashList.second[iflash];
+	            if(flash->TotalPE()<fFlashPeThresh){
+	                continue;
+	            }
+
+	            double tflash = flash->Time();//*1e3;//-fOpDelay;
+
+	            TVector3 rflash(0,flash->YCenter(),flash->ZCenter());
+	            TVector3 vdiff = rcrt-rflash;
+	            //std::cout << "flash time: "<<flash->Time() << " absdiff : "<< abs(tcrt-tflash)<< std::endl;
+	            if(abs(tcrt-tflash)<abs(tdiff)) 
+                {
+	                peflash = flash->TotalPE();
+	                tdiff = tcrt-tflash;
+	                rdiff = vdiff.Mag();
+	                xyzt.clear();
+	                xyzt.push_back(rflash.X());
+	                xyzt.push_back(rflash.Y());
+	                xyzt.push_back(rflash.Z());
+	                xyzt.push_back(tflash);
+	                matched = true;
+	                matchtpc = flashList.first;
+
+	                vector<art::Ptr<recob::OpHit>> hits = findManyHits.at(iflash);
+	                for(auto const& hit : hits) 
+                    {
+	                    double tPmt = hit->PeakTime();//*1.e3;//-fOpDelay;
+	                    if( tPmt < flashHitT) 
+                        {
+	                        flashHitT = tPmt;
+	                        flashHitPE = hit->PE();
+	                        //FlashHit position/time
+	                        double pos[3];
+	                        fGeometryService->OpDetGeoFromOpChannel(hit->OpChannel()).GetCenter(pos);
+	                        flashHitxyzt.clear();
+	                        for(int i=0; i<3; i++) flashHitxyzt.push_back(pos[i]);
+	                        flashHitxyzt.push_back(flashHitT);
+
+	                        //FlashHit distance
+	                        TVector3 rflashHit(pos[0],pos[1],pos[2]);
+	                        TVector3 vdiffHit = rcrt-rflashHit;
+	                        flashHitDiff = vdiffHit.Mag();
+	                    }
+	                }//loop over flash hits
+	            }//if minimum tdiff
+            }//for OpFlash in this flash list
+        }//for flash lists
+
+        if(!matched) {
+            peflash = DBL_MAX;
+            for(int i=0; i<4; i++) xyzt.push_back(DBL_MAX);
+        }
+
+        fMatchFlash.push_back(matched);
+        fTofFlash.push_back(tdiff);
+        fTofPeFlash.push_back(peflash);
+        fTofXYZTFlash.push_back(xyzt);
+        fDistFlash.push_back(rdiff);
+        fTofTpcFlash.push_back(matchtpc);
+        fTofFlashHit.push_back(tcrt-flashHitT);
+        fTofPeFlashHit.push_back(flashHitPE);
+        fTofXYZTFlashHit.push_back(flashHitxyzt);
+        fDistFlashHit.push_back(flashHitDiff);
+
+        // -- match OpHits to CRTHits --
+        tdiff = DBL_MAX;
+        matched = false;
+        peflash = DBL_MAX;
+        rdiff = DBL_MAX;
+        xyzt.clear();
+
+        double pemax = 0.;
+        for(auto const& hit : opHitList) 
+        {
+            double thit = hit->PeakTime();//*1e3-fOpDelay;
+            if(hit->PE()<fHitPeThresh){
+	            continue;
+            }
+
+            //if(abs(tcrt-thit)<abs(tdiff)) {
+            if(abs(tcrt-thit)<fCoinWindow && hit->PE()>pemax) 
+            {
+	            pemax = hit->PE();
+
+	            //hitXYZT
+	            double pos[3];
+	            fGeometryService->OpDetGeoFromOpChannel(hit->OpChannel()).GetCenter(pos);
+
+	            //distHit
+	            TVector3 rhit (pos[0],pos[1],pos[2]);
+	            TVector3 vdiff = rcrt-rhit;
+	            rdiff = vdiff.Mag();
+	            peflash = hit->PE();
+	            tdiff = tcrt-thit;
+	            xyzt.clear();
+	            for(int i=0; i<3; i++) xyzt.push_back(pos[i]);
+	            xyzt.push_back(thit);
+
+	            matched = true;
+	            std::cout << "thit: "<<thit << " , tcrt: " << tcrt << " , tdiff " << tdiff  << std::endl;
+            }//if min tof
+        }//for OpHits
+
+        if(!matched) {
+            tdiff = DBL_MAX;
+            peflash = DBL_MAX;
+            rdiff = DBL_MAX;
+            xyzt.clear();
+            for(int i=0; i<4; i++) xyzt.push_back(DBL_MAX);
+        }
+
+        fMatchHit.push_back(matched);
+        fTofHit.push_back(tdiff);
+        fTofPeHit.push_back(peflash);
+        fTofXYZTHit.push_back(xyzt);
+        fDistHit.push_back(rdiff);
+        fTofTpcHit.push_back(matchtpc);
+    }//for CRTHits
+
+    fMatchTree->Fill();
 }
 
 void icarus::crt::CRTPMTMatchingAna::ClearVecs()
 {
-  //matchTree
-  fCrtXYZT.clear();
-  fCrtXYZErr.clear();
-  fCrtRegion.clear();
-  fCrtPE.clear();
-  fTofHit.clear();
-  fTofFlash.clear();
-  fTofFlashHit.clear();
-  fDistHit.clear();
-  fDistFlash.clear();
-  fDistFlashHit.clear();
-  fTofPeHit.clear();
-  fTofPeFlash.clear();
-  fTofPeFlashHit.clear();
-  fTofXYZTHit.clear();
-  fTofXYZTFlash.clear();
-  fTofXYZTFlashHit.clear();
-  fTofTpcHit.clear();
-  fTofTpcFlash.clear();
-  fMatchHit.clear();
-  fMatchFlash.clear();
+    //matchTree
+    fCrtXYZT.clear();
+    fCrtXYZErr.clear();
+    fCrtRegion.clear();
+    fCrtPE.clear();
+    fTofHit.clear();
+    fTofFlash.clear();
+    fTofFlashHit.clear();
+    fDistHit.clear();
+    fDistFlash.clear();
+    fDistFlashHit.clear();
+    fTofPeHit.clear();
+    fTofPeFlash.clear();
+    fTofPeFlashHit.clear();
+    fTofXYZTHit.clear();
+    fTofXYZTFlash.clear();
+    fTofXYZTFlashHit.clear();
+    fTofTpcHit.clear();
+    fTofTpcFlash.clear();
+    fMatchHit.clear();
+    fMatchFlash.clear();
 }
 
 bool icarus::crt::CRTPMTMatchingAna::HitCompare(const art::Ptr<CRTHit>& hit1, const art::Ptr<CRTHit>& hit2) {
 
-  if(hit1->ts1_ns != hit2->ts1_ns) return false;
-  if(hit1->plane  != hit2->plane) return false;
-  if(hit1->x_pos  != hit2->x_pos) return false;
-  if(hit1->y_pos  != hit2->y_pos) return false;
-  if(hit1->z_pos  != hit2->z_pos) return false;
-  if(hit1->x_err  != hit2->x_err) return false;
-  if(hit1->y_err  != hit2->y_err) return false;
-  if(hit1->z_err  != hit2->z_err) return false;
-  if(hit1->tagger != hit2->tagger) return false;
-
-  return true;
+    if(hit1->ts1_ns != hit2->ts1_ns) return false;
+    if(hit1->plane  != hit2->plane) return false;
+    if(hit1->x_pos  != hit2->x_pos) return false;
+    if(hit1->y_pos  != hit2->y_pos) return false;
+    if(hit1->z_pos  != hit2->z_pos) return false;
+    if(hit1->x_err  != hit2->x_err) return false;
+    if(hit1->y_err  != hit2->y_err) return false;
+    if(hit1->z_err  != hit2->z_err) return false;
+    if(hit1->tagger != hit2->tagger) return false;
+    
+    return true;
 }
 
 DEFINE_ART_MODULE(icarus::crt::CRTPMTMatchingAna)
