+++ icaruscode/Decode/CMakeLists.txt
@@ -48,7 +48,7 @@ simple_plugin(DaqDecoderICARUSPMT module
   icaruscode_Decode_DecoderTools_Dumpers
   icaruscode_Decode_DecoderTools
   icaruscode_Timing
-  icaruscode_Timing_DataProducts
+  icaruscode_IcarusObj
   sbndaq_artdaq_core::sbndaq-artdaq-core_Overlays_Common
   sbndaq_artdaq_core::sbndaq-artdaq-core_Overlays
   lardataobj_RawData
