+++ icaruscode/Decode/fcl/Definitions/CAEN_V1730_setup_icarus.fcl
@@ -119,6 +119,7 @@ icarus_V1730_setup_Run0: [
 icarus_V1730_16thChannel_trgprim_setup_Run1: {
   ChannelIndex:          15  # the last one
   Channel:             @nil  # needs to be customized on each board
+  OnlyOnGlobalTrigger:   true
   InstanceName:          "trgprim"
 }
 icarus_V1730_16thChannel_RWM_setup_Run1: {
