+++ icaruscode/Decode/fcl/Definitions/decoderdefs_icarus.fcl
@@ -50,6 +50,7 @@ decodePMT: {
                     PMTconfigTag:       @nil # must override
                     TriggerTag:         @nil # must override
                     BoardSetup:         @local::icarus_V1730_setup
+                    CorrectionInstance: "trgprim"
 }
 
 decodeTrigger: {
