+++ icaruscode/IcarusObj/CMakeLists.txt
@@ -1,10 +1,14 @@
 
-cet_make_library(SOURCE 
-                 ChannelROI.cxx
-                 LIBRARIES
-                 PRIVATE
-                 messagefacility::MF_MessageLogger
-                )
+##cet_make_library(SOURCE 
+##                 ChannelROI.cxx
+##                 PMTWaveformTimeCorrection.cxx
+##                 LIBRARIES
+##                 PRIVATE
+##                 messagefacility::MF_MessageLogger
+##                )
+cet_make (
+           NO_DICTIONARY
+         )
 
 art_dictionary(DICTIONARY_LIBRARIES
                  icaruscode::IcarusObj
