+++ icaruscode/IcarusObj/classes.h
@@ -1,15 +1,20 @@
 #include "canvas/Persistency/Common/Wrapper.h"
 #include "canvas/Persistency/Common/Ptr.h"
 #include "canvas/Persistency/Common/Assns.h"
+
 #include "icaruscode/IcarusObj/SimEnergyDepositSummary.h"
 #include "icaruscode/IcarusObj/OpDetWaveformMeta.h"
+#include "icaruscode/IcarusObj/PMTWaveformTimeCorrection.h"
 #include "icaruscode/IcarusObj/ChannelROI.h"
+
 #include "sbnobj/ICARUS/PMT/Trigger/Data/OpticalTriggerGate.h"
 #include "sbnobj/ICARUS/PMT/Trigger/Data/TriggerGateData.h"
+
 #include "lardataobj/AnalysisBase/T0.h"
 #include "lardataobj/RecoBase/PFParticle.h"
 #include "lardataobj/Simulation/BeamGateInfo.h"
 #include "lardataobj/RawData/OpDetWaveform.h"
+
 #include <vector>
 
 namespace {
