--- icaruscode/Timing/DataProducts/CMakeLists.txt
@@ -1,8 +1,0 @@
-cet_make( 
-  NO_DICTIONARY
-  )
-
-art_dictionary(DICTIONARY_LIBRARIES icaruscode_Timing_DataProducts)
-
-install_headers()
-install_source()
