--- icaruscode/Timing/DataProducts/PMTWaveformTimeCorrection.cxx
@@ -1,1 +1,0 @@
-#include "PMTWaveformTimeCorrection.h"
