--- icaruscode/Timing/DataProducts/classes.h
@@ -1,17 +1,0 @@
-/**
- * @file   icaruscode/Timing/DataProducts/classes.h
- * @author Andrea Scarpelli (ascarpell@bnl.gov)
- * @date   September 1, 2022
- * 
- * Enables dictionary definitions for:
- * 
- * * `icarus::timing::PMTWaveformTimeCorrection"`
- * 
- * See also `icaruscode/Timing/DataProducts/classes_def.xml`.
- */
-
-#include "icaruscode/Timing/DataProducts/PMTWaveformTimeCorrection.h"
-
-// framework libraries
-#include "canvas/Persistency/Common/Ptr.h"
-#include "canvas/Persistency/Common/Wrapper.h"
